Id, int to long

+++ src/main/java/com/backend/domain/bid/dto/BidResponseDto.java
@@ -3,10 +3,10 @@ package com.backend.domain.bid.dto;
 import java.time.LocalDateTime;
 
 public record BidResponseDto (
-  Integer id,
-  Integer productId,
-  Integer bidderId,
-  long price,
-  String status,
-  LocalDateTime createDate
+        Long id,
+        Long productId,
+        Long bidderId,
+        long price,
+        String status,
+        LocalDateTime createDate
 ) {}
